requirements.txt
Flask==1.1.1
jaconv==0.2.4
pykakasi==1.2
google-cloud-texttospeech==0.5.0